cleanup visibility, turn on visibility lints
